requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
python-dateutil==2.8.2
feedparser==6.0.10
lxml==4.9.3
Werkzeug==3.0.1
gunicorn==21.2.0
